refactor: Move shared task runner types to task runner package
