+++ packages/@n8n/task-runner/src/index.ts
@@ -1,2 +1,3 @@
 export * from './task-runner';
 export * from './runner-types';
+export * from './runner-messages';
