+++ packages/cli/src/runners/task-managers/task-manager.ts
@@ -1,92 +1,30 @@
-import {
-	type EnvProviderState,
-	type IExecuteFunctions,
-	type Workflow,
-	type IRunExecutionData,
-	type INodeExecutionData,
-	type ITaskDataConnections,
-	type INode,
-	type WorkflowParameters,
-	type INodeParameters,
-	type WorkflowExecuteMode,
-	type IExecuteData,
-	type IDataObject,
-	type IWorkflowExecuteAdditionalData,
-	type Result,
-	createResultOk,
-	createResultError,
+import type { TaskResultData, RequesterMessage, N8nMessage, TaskData } from '@n8n/task-runner';
+import { RPC_ALLOW_LIST } from '@n8n/task-runner';
+import type {
+	EnvProviderState,
+	IExecuteFunctions,
+	Workflow,
+	IRunExecutionData,
+	INodeExecutionData,
+	ITaskDataConnections,
+	INode,
+	INodeParameters,
+	WorkflowExecuteMode,
+	IExecuteData,
+	IDataObject,
+	IWorkflowExecuteAdditionalData,
+	Result,
 } from 'n8n-workflow';
+import { createResultOk, createResultError } from 'n8n-workflow';
 import { nanoid } from 'nanoid';
 
 import { DataRequestResponseBuilder } from './data-request-response-builder';
-import {
-	RPC_ALLOW_LIST,
-	type TaskResultData,
-	type N8nMessage,
-	type RequesterMessage,
-} from '../runner-types';
 
 export type RequestAccept = (jobId: string) => void;
 export type RequestReject = (reason: string) => void;
 
 export type TaskAccept = (data: TaskResultData) => void;
 export type TaskReject = (error: unknown) => void;
-
-export interface TaskData {
-	executeFunctions: IExecuteFunctions;
-	inputData: ITaskDataConnections;
-	node: INode;
-
-	workflow: Workflow;
-	runExecutionData: IRunExecutionData;
-	runIndex: number;
-	itemIndex: number;
-	activeNodeName: string;
-	connectionInputData: INodeExecutionData[];
-	siblingParameters: INodeParameters;
-	mode: WorkflowExecuteMode;
-	envProviderState: EnvProviderState;
-	executeData?: IExecuteData;
-	defaultReturnRunIndex: number;
-	selfData: IDataObject;
-	contextNodeName: string;
-	additionalData: IWorkflowExecuteAdditionalData;
-}
-
-export interface PartialAdditionalData {
-	executionId?: string;
-	restartExecutionId?: string;
-	restApiUrl: string;
-	instanceBaseUrl: string;
-	formWaitingBaseUrl: string;
-	webhookBaseUrl: string;
-	webhookWaitingBaseUrl: string;
-	webhookTestBaseUrl: string;
-	currentNodeParameters?: INodeParameters;
-	executionTimeoutTimestamp?: number;
-	userId?: string;
-	variables: IDataObject;
-}
-
-export interface DataRequestResponse {
-	workflow: Omit<WorkflowParameters, 'nodeTypes'>;
-	inputData: ITaskDataConnections;
-	node: INode;
-
-	runExecutionData: IRunExecutionData;
-	runIndex: number;
-	itemIndex: number;
-	activeNodeName: string;
-	connectionInputData: INodeExecutionData[];
-	siblingParameters: INodeParameters;
-	mode: WorkflowExecuteMode;
-	envProviderState: EnvProviderState;
-	executeData?: IExecuteData;
-	defaultReturnRunIndex: number;
-	selfData: IDataObject;
-	contextNodeName: string;
-	additionalData: PartialAdditionalData;
-}
 
 export interface TaskRequest {
 	requestId: string;
